CI: Run tests without ASLR (#162)
The boot images are built with the assumption that they will be loaded
at a specific address. ASLR will cause some shared objects to be loaded
at the addresses of the boot images, and the boot images will overwrite
their sections, causing illegal instruction errors and other errors.
This problem is easy to reproduce locally, and manifests very frequently
when running on GitHub CI.

This PR modifies CI scripts to mitigate the problem. It adds `setarch
-R` before executing `rvm` to disable ASLR.

This PR is a workaround. The proper fix is take control over the loading
process and not let the system loader load shared objects in the address
range of the boot image.

Related issue: https://github.com/mmtk/mmtk-jikesrvm/issues/158